pyproject.toml
[tool.poetry]
name = "my-first-ml-pipeline"
version = "0.1.0"
description = ""
authors = ["Lev Plotkin <{ID}+{username}@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.1"
matplotlib = "^3.8.3"
scikit-learn = "^1.4.1.post1"
nyoka = "^5.5.0"
pypmml = "^0.9.17"
sklearn-pmml-model = "^1.0.6"
skompiler = "^0.7"
titus = "^0.8.4"
gitpython = "^3.1.42"
mlflow = "^2.11.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

